added scoringbeta and flagdata.
scoringbeta might work. flagdata needs to be updated with separate flags, so we need to find a way to work around that. not sure how at the moment. we might just have to tell ken to rotate the box every week and update flagdata since right now it only supports 1 flag